redme.md file added

+++ src/MyCraft/Mycrafts.jsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../Provider/AuthProvider";
 import { Link } from "react-router-dom";
-
+import Swal from 'sweetalert2'
 
 const Mycrafts = () => {
     const {user} = useContext(AuthContext)
@@ -11,23 +11,41 @@ const Mycrafts = () => {
     // { _id, photo, name, subcategory, time, price, rating, 
     // customization, stockStatus, description, userName, userEmail }
     useEffect(()=>{
-        fetch(`http://localhost:5000/mycrafts/${user?.email}`)
+        fetch(`https://paper-crafts-and-glass-art-server.vercel.app/mycrafts/${user?.email}`)
         .then(res=>res.json())
         .then(data=>{
             setCrafts(data);
         })
     },[user, control])
 const handleDelete = (id)=>{
-    fetch(`http://localhost:5000/delete/${id}`,{
-        method:'DELETE'
-    })
-    .then(res=>res.json())
-    .then(data=>{
-        console.log(data);
-        if(data.deletedCount>0){
-            setControl(!control)
-        }
-    })
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://paper-crafts-and-glass-art-server.vercel.app/delete/${id}`,{
+                    method:'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        if (data.deletedCount > 0) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your Craft has been deleted.',
+                                'success'
+                            )
+                            setControl(!control)
+                        }
+                    })
+
+            }
+        })
 }
     return (
         <div className="py-20 container mx-auto">
@@ -56,8 +74,8 @@ const handleDelete = (id)=>{
                         </div>
                         <hr />
                         <div className="flex justify-end gap-5">
-                            <Link to={`/updateCraft/${craft._id}`} className="btn">Update</Link>
-                            <button onClick={()=>handleDelete(craft._id)} className="btn">Delete</button>
+                            <Link to={`/updateCraft/${craft._id}`} className="btn bg-[#ff6b6b] text-white">Update</Link>
+                            <button onClick={()=>handleDelete(craft._id)} className="btn bg-red-700 text-white">Delete</button>
                         </div>
                     </div>
                 </div>
